pyproject.toml
[project]
name = "email-summarizer"
version = "0.1.0"
description = ""
authors = [
    {name = "David Harmon",email = "dharmon202@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "boto3 (>=1.37.33,<2.0.0)",
    "google-api-python-client (>=2.167.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pystache (>=0.6.8,<0.7.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
]

[tool.poetry]
packages = [{include = "email_summarizer", from = "src"}]


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
pre-commit = "^4.2.0"
black = "^25.1.0"
isort = "^6.0.1"
types-boto3 = "^1.0.1"
mypy = "^1.15.0"
tox = "^4.25.0"


[tool.poetry.group.test.dependencies]
tox = "^4.25.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
